+++ packages/gamut/src/Tip/PreviewTip/utils.ts
@@ -1,0 +1,11 @@
+import { PreviewTipProps } from '.';
+
+export const getPreviewDescription = ({
+  linkDescription,
+  overline,
+  username,
+}: Pick<PreviewTipProps, 'linkDescription' | 'overline' | 'username'>) => {
+  return `Preview: ${overline ? `${overline} ` : ''}${
+    username ? `${username} ` : ''
+  }${linkDescription}`;
+};