+++ packages/gamut/src/Tip/index.tsx
@@ -1,4 +1,0 @@
-export * from './DeprecatedToolTip';
-export * from './InfoTip';
-export * from './ToolTip';
-export type { DeprecatedToolTipProps } from './DeprecatedToolTip/types';