+++ packages/gamut-patterns/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.9.15](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.9.14...@codecademy/gamut-patterns@0.9.15) (2024-10-01)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
 ### [0.9.14](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-patterns@0.9.13...@codecademy/gamut-patterns@0.9.14) (2024-08-27)
 
 **Note:** Version bump only for package @codecademy/gamut-patterns